requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.28
email-validator==2.1.0.post1
fpdf==1.7.2
schedule==1.2.1
Flask-Limiter==3.5.1
Flask-Caching==2.1.0
Flask-Compress==1.14
Flask-Talisman==1.1.0
redis==5.0.1
cryptography==42.0.2
bcrypt==4.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
colorama==0.4.6
blinker==1.7.0
limits==3.7.0
WTForms==3.1.2
alembic==1.13.1 